requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pyyaml
pytest
httpx
numpy
scikit-learn
matplotlib